+++ bot/commands/.__init__.py
@@ -1,0 +1,5 @@
+from .character import Character
+from .creature import Creature
+from .location import Location
+from .spell import Spell
+from .admin import Admin
