+++ bot/commands/location.py
@@ -1,0 +1,55 @@
+"""Commands that refer to a location.
+
+- !create_settlement
+
+"""
+import discord
+from discord.ext import commands
+from psycopg2.extensions import connection
+
+class Location(commands.Cog):
+    """commands that manage a location"""
+
+    def __init__(self, bot: commands.bot.Bot, conn: connection):
+        self.bot = bot
+        self.conn = conn
+    
+    def is_valid_settlement(self, ctx) -> tuple[bool, str]:
+        """
+        Checks if the command location is a valid settlement
+        i.e. A thread inside of a forum
+        """
+        if not isinstance(ctx.channel, discord.Thread):
+            return False, "This command must be used inside a forum thread."
+
+        if not isinstance(ctx.channel.parent, discord.ForumChannel):
+            return False, "This thread is not part of a forum channel."
+        return True, None
+
+    @commands.command()
+    async def create_settlement(self, ctx):
+        """Sets a channel as a location"""
+        if not ctx.author.guild_permissions.administrator:
+            await ctx.send("You must be an admin to use this command.")
+            return
+        
+        is_valid, message = self.is_valid_settlement(ctx)
+        if not is_valid:
+            await ctx.send(message)
+            return
+
+        await ctx.send(
+            f"Forum Channel: **{ctx.channel.parent.name}** (ID: `{ctx.channel.parent.id}`)\n"
+            f"Thread: **{ctx.channel.name}** (ID: `{ctx.channel.id}`)"
+        )
+
+        location_map = get_location_mapping(conn, ctx.guild.id)
+        if ctx.channel.parent.id not in location_map.keys():
+            print(
+                f"https://discord.com/channels/{ctx.guild.id}/{ctx.channel.parent.id}")
+            await ctx.send(generate_location(ctx, conn))
+            print(f"Generated Location {ctx.channel.parent.name}")
+        location_map = get_location_mapping(conn, ctx.guild.id)
+        settlement_map = get_settlement_mapping(conn, ctx.guild.id)
+        if ctx.channel.id not in settlement_map.keys():
+            await ctx.send(generate_settlement(ctx, conn, location_map))
